+++ client/src/context/search/index.ts
@@ -1,3 +1,2 @@
-export * from './search';
-export * from './search.constants';
-export { useActiveQueryParameter } from './search.hooks';
+export * from './constants';
+export * from './search.context';
