+++ client/src/context/search/search.context.tsx
@@ -1,0 +1,46 @@
+import { createContext, useContext } from 'react';
+import { ReactNode } from 'react-markdown';
+
+import { PluginIndexData } from '@/types';
+
+import { useFilters } from './filter.hooks';
+import { useSearch } from './search.hooks';
+import { useSort } from './sort.hooks';
+import { SearchState } from './types';
+
+interface Props {
+  children: ReactNode;
+  pluginIndex: PluginIndexData[];
+}
+
+const PluginSearchStateContext = createContext<SearchState | null>(null);
+
+/**
+ * Hook for accessing the search state context.
+ */
+export function useSearchState(): SearchState | null {
+  return useContext(PluginSearchStateContext);
+}
+
+/**
+ * Provider for search state.  This allows child components to get/set the query
+ * and get search results.
+ */
+export function PluginSearchProvider({ children, pluginIndex }: Props) {
+  const { results, searchForm } = useSearch(pluginIndex);
+  const { filterForm, filteredResults } = useFilters(results);
+  const { sortForm, sortedResults } = useSort(filteredResults);
+
+  return (
+    <PluginSearchStateContext.Provider
+      value={{
+        filter: filterForm,
+        results: sortedResults,
+        search: searchForm,
+        sort: sortForm,
+      }}
+    >
+      {children}
+    </PluginSearchStateContext.Provider>
+  );
+}
